Enhancement #1
Address enhancement #1. You can now upload a JSON file with a roster
and view the contents.  Had to add unique keys for each player when
processing JSON file, and had to normalize name output.  For now to
view a different lineup you need to refresh the page.  Not ideal, but
quick and dirty.